chore: move apllyNewMCDAConfig to separate file to remove circular dependency. mcda/utils folder
