+++ src/features/mcda/index.ts
@@ -3,18 +3,14 @@ import { toolbar } from '~core/toolbar';
 import { store } from '~core/store/store';
 import { enabledLayersAtom } from '~core/logical_layers/atoms/enabledLayers';
 import { getMutualExcludedActions } from '~core/logical_layers/utils/getMutualExcludedActions';
-import { layersSourcesAtom } from '~core/logical_layers/atoms/layersSources';
-import { applyNewLayerSource } from '~core/logical_layers/utils/applyNewLayerSource';
-import { deepCopy } from '~core/logical_layers/utils/deepCopy';
 import { mcdaLayerAtom } from './atoms/mcdaLayer';
 import { createMCDAConfig, editMCDAConfig } from './mcdaConfig';
 import { MCDA_CONTROL_ID, UPLOAD_MCDA_CONTROL_ID } from './constants';
-import { askMcdaJSONFile } from './openMcdaFile';
+import { askMcdaJSONFile } from './utils/openMcdaFile';
 import type {
   LogicalLayerActions,
   LogicalLayerState,
 } from '~core/logical_layers/types/logicalLayer';
-import type { MCDAConfig } from '~core/logical_layers/renderers/stylesConfigs/mcda/types';
 
 export const mcdaControl = toolbar.setupControl({
   id: MCDA_CONTROL_ID,
@@ -77,18 +73,6 @@ uploadMcdaControl.onStateChange((ctx, state) => {
     store.dispatch(uploadMcdaControl.setState('regular'));
   }
 });
-
-export function applyNewMCDAConfig(config: MCDAConfig) {
-  const id = config.id;
-  const oldSource = layersSourcesAtom.getState().get(id)?.data;
-  if (oldSource) {
-    const newSource = deepCopy(oldSource);
-    if (newSource?.style?.config) {
-      newSource.style.config = { ...config };
-    }
-    applyNewLayerSource(newSource);
-  }
-}
 
 /** Edit MCDA */
 export async function editMCDA(
